+++ EduVPN-multiOS/Coordinators/AppCoordinator Delegates/VPNConnectionViewControllerDelegate.swift
@@ -23,8 +23,8 @@ protocol VPNConnectionViewControllerDelegate: class {
 
 extension AppCoordinator: VPNConnectionViewControllerDelegate {
     func confirmDisconnectWhileOnDemandEnabled() -> Promise<Bool> {
-        return showActionSheet(title: NSLocalizedString("OnDemand enabled", comment: ""),
-                               message: NSLocalizedString("Are you sure you want to disconnect while OnDemand is enabled?", comment: ""),
+        return showActionSheet(title: NSLocalizedString("On Demand enabled", comment: ""),
+                               message: NSLocalizedString("Are you sure you want to disconnect while “On Demand” is enabled?", comment: ""),
                                confirmTitle: NSLocalizedString("Disconnect", comment: ""),
                                declineTitle: NSLocalizedString("Cancel", comment: ""))
     }